feat(prepend): prepend condition from env `WRAPPED_PREPEND_IF` which use a regex to match args list
We add a new env `WRAPPED_PREPEND_IF`. It can be a regex, and is matched with the origin arguments
list. If at least one argument matches the regex, the prepending condition will be set to true and
cmd-wrapper will prepend arguments specified by env `WRAPPED_PREPEND_ARGS` to the new arguments
list.
Note, as a special case, if env `WRAPPED_PREPEND_IF` is empty, the prepending condition would
always be set to true.